Centralize/Expand Setup.
Attempt to better encompass the entire world of setup around the
application. Node and GIT can be common place for most of us but, for a
lot of people they are going to be coming into the project a little
blind. Make super explicit (try) the steps to setup the repo +
development tools to make someone as successful as possible.