Resolve Issue #25: Nav overlay content instead of pushing down (#44)
* Testing Now deployment

* additional bulma changes

* #4 Options for footer display (#12)

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* increave viewport over 100%/vh to force scroll

* Create pages with filler text



* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* increave viewport over 100%/vh to force scroll

* remove extraneous code added

* nav bar links and initial link pages

* remove invalid character

* run npm audit

* Update README with instructions for running project

* Update README with instructions for running project

* #10 Join Us button changes (#14)

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* increave viewport over 100%/vh to force scroll

* remove extraneous code added

* nav bar links and initial link pages

* remove invalid character

* join us button updates

* Join Us button fixes

* Migrate Lorem IPSUM In About Page To Old Content. (#15)

* Migrate Lorem IPSUM To Old Content.

Merge in content from the older code for greensobor RAILS site. Some
link to helpers sticking around but for the most part.

Co-authored-by: John Hill <john.m.hill91@gmail.com>

* Move Link To Helpers To Actual A Tags.

Next has some wonderful helper link tags but, none of these URL's are
internal to the application.

Co-authored-by: John Hill <john.m.hill91@gmail.com>

* Move To ClassName Over Class

The copy pasta of the older rails version does support the actual dom
API. Good for you React. Move to ClassName over Class to make the next
CLI happy.

* Joinuspage (#16)

* played with

* worked on join us page

* #3Fix logo (#17)

* Code pic logo .

* Modify brand banner.

* #3 Logo component and standardize classNames/styles (#18)

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* 2 options for footer: sticky bottom or end content

* increave viewport over 100%/vh to force scroll

* remove extraneous code added

* nav bar links and initial link pages

* remove invalid character

* join us button updates

* Join Us button fixes

* logo component (allow reuse), standardize styles

* Move The Next Readme Into The Doc Folder.

Begin centralizing documentation in a docs folder. Readmes have a
tendency to get out of hand and are hard to follow and a bit of a bear
to update. Create a new docs folder to break things apart and give
people bite sized chunks of markdown files.

* Centralize/Expand Setup.

Attempt to better encompass the entire world of setup around the
application. Node and GIT can be common place for most of us but, for a
lot of people they are going to be coming into the project a little
blind. Make super explicit (try) the steps to setup the repo +
development tools to make someone as successful as possible.

* Root Readme As A Table Of Contents

The root readme is still going to be what's showcased front and center
when someone navigates to the GitHub page. Turn this into a table of
contents so folks can easily navigate their way to the relevant part of
the doc folder.

* Update fontawesome. Add content to index.js

* FA updates

* Removed pages/resources.js and removed link to the resources page form components/nav.js

* Changes to setup, scss, and projects (#30)

* updated setup.md for more clarity in new user setup

* Added h2 to styles and started adding projects.

* Add style and change DOM element (#31)

* Fix error caused by a blockquote inside a p tag on About page (#29)

* Fix error caused by a blockquote inside a p tag on About page

* Move the blockquote outside the paragraph

* Added image to home page & about page (#33)

* Added image to home page & about page

* Removing webpack loader

* Using png format for image

* Implement eslint/prettier (#35)

* Implement eslint/prettier

* Fix next config

* Issue #25: navbar overlay